put mc_comm and mc_simulator tests in theirown folders and renamed fifo simulator folder
