+++ tests/mc_comm/test_mc_comm.py
@@ -31,17 +31,17 @@ from stdlib_utils import drain_queue
 from stdlib_utils import InfiniteProcess
 from stdlib_utils import invoke_process_run_and_check_errors
 
-from .fixtures import generate_board_and_error_queues
-from .fixtures import QUEUE_CHECK_TIMEOUT_SECONDS
-from .fixtures_mc_comm import fixture_four_board_mc_comm_process
-from .fixtures_mc_comm import fixture_patch_comports
-from .fixtures_mc_comm import fixture_patch_serial_connection
-from .fixtures_mc_simulator import fixture_mantarray_mc_simulator
-from .fixtures_mc_simulator import fixture_mantarray_mc_simulator_no_beacon
-from .helpers import assert_queue_is_eventually_not_empty
-from .helpers import confirm_queue_is_eventually_empty
-from .helpers import confirm_queue_is_eventually_of_size
-from .helpers import put_object_into_queue_and_raise_error_if_eventually_still_empty
+from ..fixtures import generate_board_and_error_queues
+from ..fixtures import QUEUE_CHECK_TIMEOUT_SECONDS
+from ..fixtures_mc_comm import fixture_four_board_mc_comm_process
+from ..fixtures_mc_comm import fixture_patch_comports
+from ..fixtures_mc_comm import fixture_patch_serial_connection
+from ..fixtures_mc_simulator import fixture_mantarray_mc_simulator
+from ..fixtures_mc_simulator import fixture_mantarray_mc_simulator_no_beacon
+from ..helpers import assert_queue_is_eventually_not_empty
+from ..helpers import confirm_queue_is_eventually_empty
+from ..helpers import confirm_queue_is_eventually_of_size
+from ..helpers import put_object_into_queue_and_raise_error_if_eventually_still_empty
 
 __fixtures__ = [
     fixture_four_board_mc_comm_process,
